refactored maintenance app into a cli and a library
